+++ misskey-api/src/endpoint/admin/emoji/add.rs
@@ -1,5 +1,7 @@
 #[cfg(feature = "12-9-0")]
 use crate::model::drive::DriveFile;
+#[cfg(feature = "13-13-0")]
+use crate::model::role::Role;
 use crate::model::{emoji::Emoji, id::Id};
 
 use serde::{Deserialize, Serialize};
@@ -14,23 +16,43 @@ pub struct Request {
     #[cfg(feature = "12-9-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "12-9-0")))]
     pub file_id: Id<DriveFile>,
-    #[cfg(any(docsrs, not(feature = "12-9-0")))]
-    #[cfg_attr(docsrs, doc(cfg(not(feature = "12-9-0"))))]
+    #[cfg(any(docsrs, not(feature = "12-9-0"), feature = "13-13-0"))]
+    #[cfg_attr(docsrs, doc(cfg(any(not(feature = "12-9-0"), feature = "13-13-0"))))]
     #[builder(setter(into))]
     pub name: String,
     #[cfg(any(docsrs, not(feature = "12-9-0")))]
     #[cfg_attr(docsrs, doc(cfg(not(feature = "12-9-0"))))]
     pub url: Url,
-    #[cfg(any(docsrs, not(feature = "12-9-0")))]
-    #[cfg_attr(docsrs, doc(cfg(not(feature = "12-9-0"))))]
+    #[cfg(any(docsrs, not(feature = "12-9-0"), feature = "13-13-0"))]
+    #[cfg_attr(docsrs, doc(cfg(any(not(feature = "12-9-0"), feature = "13-13-0"))))]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option, into))]
     pub category: Option<String>,
-    #[cfg(any(docsrs, not(feature = "12-9-0")))]
-    #[cfg_attr(docsrs, doc(cfg(not(feature = "12-9-0"))))]
+    #[cfg(any(docsrs, not(feature = "12-9-0"), feature = "13-13-0"))]
+    #[cfg_attr(docsrs, doc(cfg(any(not(feature = "12-9-0"), feature = "13-13-0"))))]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub aliases: Option<Vec<String>>,
+    #[cfg(feature = "13-13-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option, into))]
+    pub license: Option<String>,
+    #[cfg(feature = "13-13-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub is_sensitive: Option<bool>,
+    #[cfg(feature = "13-13-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub local_only: Option<bool>,
+    #[cfg(feature = "13-13-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub role_ids_that_can_be_used_this_emoji_as_reaction: Option<Vec<Id<Role>>>,
 }
 
 #[derive(Deserialize, Debug, Clone)]
@@ -49,11 +71,11 @@ mod tests {
     use super::Request;
     use crate::test::{ClientExt, TestClient};
 
-    #[cfg(not(feature = "12-9-0"))]
+    #[cfg(any(not(feature = "12-9-0"), feature = "13-13-0"))]
     use ulid_crate::Ulid;
 
     #[tokio::test]
-    #[cfg(feature = "12-9-0")]
+    #[cfg(all(feature = "12-9-0", not(feature = "13-13-0")))]
     async fn request() {
         let client = TestClient::new();
         let image_url = client.avatar_url().await;
@@ -93,6 +115,56 @@ mod tests {
                 url: image_url,
                 category: Some("nice".to_string()),
                 aliases: Some(vec!["test2".to_string()]),
+            })
+            .await;
+    }
+
+    #[tokio::test]
+    #[cfg(feature = "13-13-0")]
+    async fn request() {
+        let client = TestClient::new();
+        let image_url = client.avatar_url().await;
+        let file = client.upload_from_url(image_url).await;
+        let name = Ulid::new().to_string();
+
+        client
+            .admin
+            .test(Request {
+                file_id: file.id,
+                name,
+                aliases: None,
+                category: None,
+                is_sensitive: None,
+                license: None,
+                local_only: None,
+                role_ids_that_can_be_used_this_emoji_as_reaction: None,
+            })
+            .await;
+    }
+
+    #[tokio::test]
+    #[cfg(feature = "13-13-0")]
+    async fn request_with_options() {
+        let client = TestClient::new();
+        let image_url = client.avatar_url().await;
+        let file = client.upload_from_url(image_url).await;
+        let name = Ulid::new().to_string();
+        let role = client
+            .admin
+            .test(crate::endpoint::admin::roles::create::Request::default())
+            .await;
+
+        client
+            .admin
+            .test(Request {
+                file_id: file.id,
+                name,
+                aliases: Some(vec!["alias".to_string()]),
+                category: Some("cat".to_string()),
+                is_sensitive: Some(true),
+                license: Some("license".to_string()),
+                local_only: Some(true),
+                role_ids_that_can_be_used_this_emoji_as_reaction: Some(vec![role.id]),
             })
             .await;
     }