+++ misskey-api/src/endpoint/users/lists/create_from_public.rs
@@ -1,22 +1,18 @@
 use crate::model::{id::Id, user_list::UserList};
 
 use serde::Serialize;
-use typed_builder::TypedBuilder;
 
-#[derive(Serialize, Debug, Clone, TypedBuilder)]
+#[derive(Serialize, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct Request {
+    /// [ 1 .. 100 ] characters
+    pub name: String,
     pub list_id: Id<UserList>,
-    #[cfg(feature = "13-13-0")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[builder(default, setter(strip_option))]
-    pub for_public: Option<bool>,
 }
 
 impl misskey_core::Request for Request {
     type Response = UserList;
-    const ENDPOINT: &'static str = "users/lists/show";
+    const ENDPOINT: &'static str = "users/lists/create-from-public";
 }
 
 #[cfg(test)]
@@ -26,25 +22,6 @@ mod tests {
 
     #[tokio::test]
     async fn request() {
-        let client = TestClient::new();
-        let list = client
-            .test(crate::endpoint::users::lists::create::Request {
-                name: "test".to_string(),
-            })
-            .await;
-
-        client
-            .test(Request {
-                list_id: list.id,
-                #[cfg(feature = "13-13-0")]
-                for_public: None,
-            })
-            .await;
-    }
-
-    #[cfg(feature = "13-13-0")]
-    #[tokio::test]
-    async fn request_with_option() {
         let client = TestClient::new();
         let list = client
             .test(crate::endpoint::users::lists::create::Request {
@@ -62,8 +39,8 @@ mod tests {
 
         client
             .test(Request {
+                name: "from public".to_string(),
                 list_id: list.id,
-                for_public: Some(true),
             })
             .await;
     }