+++ misskey-api/src/streaming/channel/local_timeline.rs
@@ -2,15 +2,22 @@ use crate::model::note::Note;
 use crate::streaming::channel::NoOutgoing;
 
 use serde::{Deserialize, Serialize};
+use typed_builder::TypedBuilder;
 
 #[derive(Deserialize, Debug, Clone)]
 #[serde(rename_all = "camelCase", tag = "type", content = "body")]
 pub enum LocalTimelineEvent {
     Note(Note),
 }
-
-#[derive(Serialize, Default, Debug, Clone)]
-pub struct Request {}
+#[derive(Serialize, Default, Debug, Clone, TypedBuilder)]
+#[serde(rename_all = "camelCase")]
+pub struct Request {
+    #[cfg(feature = "13-13-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-13-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub with_replies: Option<bool>,
+}
 
 impl misskey_core::streaming::ConnectChannelRequest for Request {
     type Incoming = LocalTimelineEvent;
@@ -43,6 +50,31 @@ mod tests {
 
         future::join(
             http_client.create_note(Some("The world is fancy!"), None, None),
+            async { stream.next().await.unwrap().unwrap() },
+        )
+        .await;
+    }
+
+    #[cfg(feature = "13-13-0")]
+    #[tokio::test]
+    async fn stream_with_replies() {
+        let http_client = HttpTestClient::new();
+        let client = TestClient::new().await;
+        let note = http_client
+            .admin
+            .create_note(Some("The world is fancy!"), None, None)
+            .await;
+        let (_, new_client) = http_client.admin.create_user().await;
+
+        let mut stream = client
+            .channel(Request {
+                with_replies: Some(true),
+            })
+            .await
+            .unwrap();
+
+        future::join(
+            new_client.create_note(Some("The world is fancy!"), None, Some(note.id)),
             async { stream.next().await.unwrap().unwrap() },
         )
         .await;